Feels like such a waste of a day, but finally have a working build environment (again), and quite a bit less mess & confusion amongst the source files
