+++ .vscode/settings.json
@@ -14,5 +14,5 @@
     "idf.pythonBinPath": "/home/chris/.espressif/python_env/idf5.1_py3.10_env/bin/python",
     "idf.toolsPath": "/home/chris/.espressif",
     "idf.flashType": "UART",
-    "sonarlint.pathToCompileCommands": "${workspaceFolder}/build/bootloader/compile_commands.json"
+    "sonarlint.pathToCompileCommands": "${workspaceFolder}/build/compile_commands.json"
 }
